% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{analyzing_means_plots}
\alias{analyzing_means_plots}
\title{analyzing_means_plots}
\usage{
analyzing_means_plots(df, num_int_var)
}
\arguments{
\item{df}{dataset to obtain the plots}

\item{num_int_var}{\emph{ONE} NUMERICAL VARIABLE from the dataset}
}
\value{
"This function returns a list where each element is going to be a sublist containing".
\itemize{
\item  \emph{bar plot} comparing the \emph{standard error} of each combination level of the factor variables for the distribution of the \code{num_int_var}.
\item \emph{Polar area plot} comparing the \emph{mean} of each combination level of the factor variables for the distribution of the \code{num_int_var}.
}
}
\description{
Create a list with specific descriptive plots analizing the mean from the distribution of a numeric variable based on the levels of factor variables from a dataset.
}
\details{
This function allows you to return descriptive plots analizing the mean from a numeric variable based on the levels of factor variables from a dataset.
}
\note{
\itemize{
\item Each output is going to be a list, where each element is the result from each combination levels of the factor variables and the whole distribution of the numeric variable.
\item If there is no factor variable(s) then it is going to bring plots analyzing only the population mean, rather than the sample means, because this function needs factor variables to compare the means of the \code{num_int_var} for the levels combinations
\item This function is applied to one numeric variable \code{num_int_var}
\item If you are not using a \emph{shiny app}, is better to use \code{general_descript_stats} function.
}
}
\author{
Eduardo Trujillo
}
