% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{assign_data_type}
\alias{assign_data_type}
\title{assign_data_type}
\usage{
assign_data_type(variable)
}
\arguments{
\item{variable}{variable that we want to assign a correct datatype}
}
\value{
This function allows us to assign the correct data type to a variable
}
\description{
Assign correct datatype to a variable
}
\details{
This function allows you to assign correctly the datatype of a variable
}
\note{
Useful when we want to assign the correct datatype to each variable of a dataset.
}
\author{
Eduardo Trujillo
}
