% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{assign_date_type}
\alias{assign_date_type}
\title{assign_date_type}
\usage{
assign_date_type(variable)
}
\arguments{
\item{variable}{variable that we want to assign date datatype}
}
\value{
This function allows us to assign the correct data type to a variable
"This function allows us to assign date data type based on:"
\itemize{
\item if a variable is date then return the variable
\item In other case return a date type date checking the possibilities (Permutation 3 to 3 of date functions)to return the date with the format ISO 8601 YYYY-MM-DD
}
}
\description{
Assign correct datatype date to a variable
}
\details{
This function allows you to assign date datatype to a variable
}
\author{
Eduardo Trujillo
}
