% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{balanced_desagregated_kmean}
\alias{balanced_desagregated_kmean}
\title{balanced_desagregated_kmean}
\usage{
balanced_desagregated_kmean(
  df,
  var_model,
  ref_desgte,
  id,
  k,
  pre_cleaning = FALSE
)
}
\arguments{
\item{df}{dataset to create the labels}

\item{var_model}{reference variable of balance}

\item{ref_desgte}{reference desagregation variable}

\item{id}{id variable reference of balance}

\item{k}{number of desire clusters}

\item{pre_cleaning}{clean or not previous to apply the k-means model}
}
\value{
dataset \code{df} with a new created variable "LABEL" with levels up
to \code{k} that balance the whole dataset where:
\itemize{
\item If \code{pre_cleaning == TRUE}, is going to clean the dataset before
applying the k-means model
\item If \code{pre_cleaning == FALSE}, is going to apply the k-means model
to the provided dataset
}
}
\description{
Create the k-means label, having in count a balanced datased based on
a unique id and a particular variable.
}
\details{
\enumerate{
\item If the user select True on the \code{pre_cleaning} then the function will
clean the dataset as preparation to obtain the k-means labels as a new
dataset variable.
\item Standarize the variable \code{var_model} according to the levels of
\code{ref_desgte}
\item Desagregate the dataset for all the observations based on the variable
\code{ref_desgte} with frequency greater to 1, taking in mind that the
\code{id} variable needs to be unique.
\item Creating the LABEL variable as the optimal k-mean labels for the dataset.
\item Optimize generated k-labels for desagregated dataset based on the
\code{id} variable.
\item Improve the dataset labels based on a balanced grouped sum of
\code{var_model} changing the labels that are upper the mean of the previous
to the lower labels.
\item Improve the dataset labels based on a balanced grouped sum of
\code{var_model} changing the labels that are out of the range population
mean +- 1 standard deviation in the same way as 6.
\item Again optmize generated k-labels for desagregated datased based on the
\code{id} variable to ensure that each \code{id} has different label.
}
}
\seealso{
\href{https://uc-r.github.io/kmeans_clustering}{k-means algorithm}
}
\author{
Eduardo Trujillo
}
