% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cardinality_adjustment}
\alias{cardinality_adjustment}
\title{cardinality_adjustment}
\usage{
cardinality_adjustment(df, k)
}
\arguments{
\item{df}{dataset to adjust cardinality}

\item{k}{number of desire clusters}
}
\value{
the dataset with the cardinality adjusted based on the desire
clusters.
}
\description{
Modify the cardinality of the dataset based on the divisibility with the
desire clusters.
}
\details{
\enumerate{
\item If the cardinality is divisible by k then return the dataset, if not then.
obtain a number called "cardinality" where under the cardinality of df is
divisible by k
\item From the elements that are not going to be in the dataset, make a
proportional sum (row cardinlaity + 1 to the last one) to add that proportion
to each numeric and integer variables on the rows of the dataset that are
going to be included.
\item Define the dataset with rows until the cardinality "cardinality"
\item Add 2. to each numeric and integer variable of the dataset to have a
balanced dataset.
}
}
\author{
Eduardo Trujillo
}
