% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{classes_vector}
\alias{classes_vector}
\title{classes_vector}
\usage{
classes_vector(data_type, df)
}
\arguments{
\item{data_type}{string datatype that we are interested}

\item{df}{dataset to use}
}
\value{
This function allows us to return vector of strings of variables that are in a certain data type
}
\description{
Bring  dataset-variable´s names from a specific datatype
}
\details{
This function allows you to brings all the variables´ names of a dataset that are in a certain data type
}
\author{
Eduardo Trujillo
}
