% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{clean_df}
\alias{clean_df}
\title{clean_df}
\usage{
clean_df(df, col_select = NULL)
}
\arguments{
\item{df}{dataset to use for cleaning}

\item{col_select}{vector of specific variables from your dataset that you want to use (rather than the all dataset).}
}
\value{
"This function allows us to \emph{clean dataset} with great characteristics:"
\itemize{
\item Dataset is in \code{data.table} format
\item Use the parameter \code{col_select}, to select variables that we want from the dataset
\item \code{Delete special characters}
\item \code{Toupper Variables}
\item \code{Toupper character and factor observations}
\item \code{Assign the correct data type to each variable}
\item \code{Change "" strings to N/A}
\item \code{Remove whitespace from start and end of each observation}
}
}
\description{
Clean a particular dataset
}
\details{
This function allows you to clean a particular dataset and if you want, only select particular variables from it.
}
\author{
Eduardo Trujillo
}
