% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{clean_merged_table}
\alias{clean_merged_table}
\title{clean_merged_table}
\usage{
clean_merged_table(df_merged, keep = c("x", "y"))
}
\arguments{
\item{df_merged}{dataset with duplicated columns}

\item{keep}{argument that indicate if you want to get .x variables or .y variables}
}
\value{
"This function returns \emph{different results} based on \code{keep} argument:"
\itemize{
\item If \code{keep = "x"} return the dataset with the no duplicated variables and the duplicated old variables without the extension .x
\item If \code{keep = "y"} return the dataset with the no duplicated variables and the duplicated new variables without the extension .y
}
}
\description{
Unify duplicated columns.
}
\details{
This function allows you after \emph{merging} datasets, unify duplicated columns (this is the extension .x and .y)
}
\note{
If there is no duplicated variables will show an string message.
}
\author{
Eduardo Trujillo
}
