% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{conditional_stats}
\alias{conditional_stats}
\title{conditional_stats}
\usage{
conditional_stats(df, num_int_var)
}
\arguments{
\item{df}{dataset to obtain the conditional information}

\item{num_int_var}{\emph{ONE} NUMERICAL VARIABLE from the dataset}
}
\value{
"This function returns a list where each element is going to be a sublist containing".
\itemize{
\item \emph{dataset with the conditional probability on that specific level-combination given the join observations on other specific level-combination} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item \emph{dataset with the conditional expectancy, conditional variance & conditional standard deviation on that specific level-combination given the join observations on other specific level-combination} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
}
}
\description{
Create a list with conditional information from a numeric variable based on the levels of factor variables from a dataset.
}
\details{
This function allows you to return conditional information from a numeric variable based on the levels of factor variables from a dataset.
}
\note{
\itemize{
\item Each output is going to be a list, where each element is the result from each combination levels of the factor variables and the whole distribution of the numeric variable.
\item If there is no factor variable(s) then it is going to bring an empty list, because this function needs factor variables to obtain the intersection of the observations of the \code{num_int_var} for the levels combinations
\item This function is applied to one numeric variable \code{num_int_var}
\item If you are not using a \emph{shiny app}, is better to use \code{general_descript_stats} function.
}
}
\author{
Eduardo Trujillo
}
