% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{delete_outliers}
\alias{delete_outliers}
\title{delete_outliers}
\usage{
delete_outliers(df, factor_vars = NULL)
}
\arguments{
\item{df}{dataset to clean outliers}

\item{factor_vars}{factor variables to delete outliers}
}
\value{
"This function allows us to return original dataset without the outliers based on:"
\itemize{
\item if \code{factor_vars = NULL} we delete every outlier of the numeric and/or integer variable without carrying about the factor variables
\item if \code{factor_vars} is not NULL then the delete of the outliers is by every level of each factor variable selected.
}
}
\description{
Delete outliers from numeric or integer variables of a dataset
}
\details{
This function allows you to delete outliers from numeric or integer variables of a dataset based on the z-score in function of the factor variables (or not)
}
\note{
The z-score indicates that when a observation has a z-score +-3 then is going to be a NA (to not affect the distribution of the variable)
}
\author{
Eduardo Trujillo
}
