% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{desagregate_df}
\alias{desagregate_df}
\title{desagregate_df}
\usage{
desagregate_df(df, ref_desgte, id)
}
\arguments{
\item{df}{original dataset to desagregate}

\item{ref_desgte}{reference desagregation variable}

\item{id}{unique id variable}
}
\value{
the dataset \code{df} desagregated by the variable \code{ref_desgte}
}
\description{
Desagregate a dateset for all the observations based on a specific variable
with frequency greater to 1, taking in mind that the id variable needs to
be unique.
}
\details{
This function has been design to check if the variable \code{id} is unique
filtered by the variable \code{ref_desgte} and if it is then for the
observations in \code{ref_desgte} with frequency greater than one desagregate
in each \code{id}, in other case return the observations in the same way for
each filter \code{id}.
}
\author{
Eduardo Trujillo
}
