% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{descr_stats}
\alias{descr_stats}
\title{descr_stats}
\usage{
descr_stats(num_int_var, marg_dist_df, marg_dist_choice = NULL)
}
\arguments{
\item{num_int_var}{\emph{ONE} NUMERICAL VARIABLE that represents the marginal distribution}

\item{marg_dist_df}{numeric or integer vector that represent a marginal distribution}

\item{marg_dist_choice}{Type of descriptive statistical analysis}
}
\value{
"This function returns \emph{different results} based on the arguments \code{marg_dist_choice} argument".
\itemize{
\item If \code{marg_dist_choice = "BOUNDARIES"}, \emph{list containing the bound, bins number, bin size and the interval values} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "DESCRIPTIVE_FREQUENCY"}, the \emph{frequency table} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "MODE_VALUE"}, the \emph{mode(s)} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "MODE_DESCRIPTION"}, the \emph{interpretation of the mode(s)} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "CENTRAL_TENDENCY"}, \emph{dataset with the mode, mean min, q25, median, q75,p90,p95,p99,max, midrange} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "DISPERSION_TENDENCY"}, \emph{dataset with the variance, standard deviation, interquantil range and range} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "COEFFICIENTS"}, \emph{list with the coefficient of variation, standard error, skeweness and kurtosis} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "COEFFICIENTS_DESCRIPTION"}, \emph{list with the interpretation of the coefficient of variation, standard error, skeweness and kurtosis} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = NULL}, list where each sublist is one of the previous mentioned.
}
}
\description{
Create a descriptive analysis of the marginal distribution of a numeric variable on a specific levels of combination from factor variables.
}
\details{
This function allows you to return different types of descriptive statistic analysis from the marginal distribution of a numeric variable on a specific levels of combination from factor variables.
}
\note{
\itemize{
\item \code{marg_dist_df} needs to be a filter sublist from \code{marg_dist} function.
\item This function is applied to one numeric variable \code{num_int_var}
\item If you are not using a \emph{shiny app}, is better to use \code{general_descript_stats} function.
}
}
\author{
Eduardo Trujillo
}
