% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{descr_stats_plots}
\alias{descr_stats_plots}
\title{descr_stats_plots}
\usage{
descr_stats_plots(num_int_var, marg_dist_df, marg_dist_choice = NULL)
}
\arguments{
\item{num_int_var}{\emph{ONE} NUMERICAL VARIABLE that represents the marginal distribution}

\item{marg_dist_df}{numeric or integer vector that represent a marginal distribution}

\item{marg_dist_choice}{Type of descriptive statistical analysis}
}
\value{
"This function returns \emph{different results} based on the arguments \code{marg_dist_choice} argument".
\itemize{
\item If \code{marg_dist_choice = "OGIVE"}, \emph{ogive plot} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "CUMMULATIVE"}, the \emph{accumulative distribution function plot} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "HISTOGRAM"}, the \emph{histogram plot} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "GENERAL_BOXPLOT"}, the \emph{box plot} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = "BINS_BOXPLOT"}, \emph{box plot on each bin from the frequency table} of the marginal distribution \code{marg_dist_df} of  \code{num_int_var} on a specific combination level.
\item If \code{marg_dist_choice = NULL}, list where each sublist is one of the previous mentioned.
}
}
\description{
Create specific descriptive plots of the marginal distribution of a numeric variable on a specific levels of combination from factor variables.
}
\details{
This function allows you to return different types of descriptive statistical plots from the marginal distribution of a numeric variable on a specific levels of combination from factor variables.
}
\note{
\itemize{
\item \code{marg_dist_df} needs to be a filter sublist from \code{marg_dist} function.
\item This function is applied to one numeric variable \code{num_int_var}
\item If you are not using a \emph{shiny app}, is better to use \code{general_descript_stats} function.
}
}
\author{
Eduardo Trujillo
}
