% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{design_DT}
\alias{design_DT}
\title{design_DT}
\usage{
design_DT(df, style = "bootstrap", extra_options = NULL, convert = FALSE, ...)
}
\arguments{
\item{df}{dataset to make the design}

\item{style}{theme of the datatable}

\item{extra_options}{\emph{list} of additional \strong{options} arguments}

\item{convert}{convert a dataset to datatable}

\item{...}{extra arguments in \code{datatable}}
}
\value{
The dataset in DT format.
"This function returns two possible results based on \code{convert} parameter:"
\itemize{
\item If \code{convert = TRUE} transform a formattable to datatable DT format
\item If \code{convert = FALSE} transform a data.table or a data.fram to datatable (DT format)
}
}
\description{
Visualize a dataset in a JS format.
}
\details{
This function allows you to transform your dataset into JS format (beautiful design).
}
\author{
Eduardo Trujillo
}
