% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{design_PLOT}
\alias{design_PLOT}
\title{design_PLOT}
\usage{
design_PLOT(
  plot,
  font_size = "italic",
  axis_color = "white",
  axis_line_color = "red"
)
}
\arguments{
\item{plot}{defined ggplot plot}

\item{font_size}{letter type of axis and title}

\item{axis_color}{color of axis}

\item{axis_line_color}{color of line axis}
}
\value{
A plot in plotly format.
}
\description{
Visualize a plot in a plotly format.
}
\details{
This function allows you to transform your plot into plotly format (beautiful design).
}
\author{
Eduardo Trujillo
}
