% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{design_report_creator}
\alias{design_report_creator}
\title{design_report_creator}
\usage{
design_report_creator(
  df,
  year = NULL,
  select_month = NULL,
  select_semester = NULL,
  select_quarter = NULL,
  summary = c("SUMMARY", "NOT_SUMMARY"),
  sparklines = FALSE
)
}
\arguments{
\item{df}{dataset to obtain the report}

\item{year}{year that we want to obtain the report}

\item{select_month}{name of the month of the year that we want to obtain the report}

\item{select_semester}{Number of the semester that we want to obtain the report 1:2}

\item{select_quarter}{Number of the quarter that we want to obtain the report 1:4}

\item{summary}{Type of report we want to obtain.}

\item{sparklines}{Type of design we want for our report}
}
\value{
"This function returns \emph{different results} based on the arguments \code{sparklines} argument".
\itemize{
\item If \code{sparklines = TRUE}, then return sparklines of a grouped dataset.
\item If \code{sparklines = FALSE}, then return the output of the function \code{report_creator} with a design for each variable
}
}
\description{
Create a periodical or historical designed-summary or designed-grouped dataset or sparklines from a grouped dataset.
}
\details{
This function allows you to return a periodical or historical designed-summary or designed-grouped dataset or sparklines for a grouped dataset from the orignal dataframe.
}
\note{
\itemize{
\item Since we are using the defined function \code{report_creator} the output is going to be the same, but now the report created will have design
\item If \code{sparklines = TRUE}, then you need to select the option code{summary = "NOT_SUMMARY"}, since it is going to create the sparklines of the distribution of a grouped dataset obtained from the function \code{report_creator}
}
}
\author{
Eduardo Trujillo
}
