% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dfsplit_test_equality}
\alias{dfsplit_test_equality}
\title{dfsplit_test_equality}
\usage{
dfsplit_test_equality(df, split_reference, test_reference, names)
}
\arguments{
\item{df}{dataset to split}

\item{split_reference}{variable of reference to split dataset}

\item{test_reference}{variable of reference to test if elements of
that variable are unique}

\item{names}{names of the sublists unique and repeated.}
}
\value{
List where each sublist is the unique and the duplicated
observations from \code{df}.
}
\description{
Split a dataset into unique and duplicate elements by an \emph{id} reference
}
\details{
This function has been design for a given dataset split it by the variable
\code{split_reference} and divide it in sublists. This is if the variable
\code{test_reference} has unique elements or not.
}
\author{
Eduardo Trujillo
}
