% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{dynamic_menuItem}
\alias{dynamic_menuItem}
\title{dynamic_menuItem}
\usage{
dynamic_menuItem(
  menu_text = NULL,
  menu_tabName = NULL,
  menu_icon = NULL,
  menu_icon_class = NULL,
  submenu_text = NULL,
  submenu_tabName = NULL,
  subitem = c("yes", "no")
)
}
\arguments{
\item{menu_text}{Show text of of the tab}

\item{menu_tabName}{id of the tab}

\item{menu_icon}{icon of the tab}

\item{menu_icon_class}{class of the icon of the tab}

\item{submenu_text}{show text of the subtabs of a tab}

\item{submenu_tabName}{id of the subtabs of a tab}

\item{subitem}{\code{yes} if there is subtabs or \code{no} if there is no subtabs}
}
\value{
"This function returns \code{two possible tabs}:"
\itemize{
\item If \code{subitem} is \code{yes} then it means that we can create a tab with subtabs
\item If \code{subitem} is \code{no} then it means that a tab doesn´t have subtabs
}
}
\description{
Create interactively the sidebar of a dashboard
}
\details{
This function allows you to create an interactive personalized tabNames and subtabnames of a dashboardPage
}
\note{
The advantage of this function is that we can create multiple tabs (thanks to the iteratior) only using this function.
}
\author{
Eduardo Trujillo
}
