% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_column_list}
\alias{filter_column_list}
\title{filter_column_list}
\usage{
filter_column_list(list, variable, element)
}
\arguments{
\item{list}{list to filter column}

\item{variable}{variable reference}

\item{element}{of the variable reference of index}
}
\value{
This function returns all the column elements based on the position of \code{element} of the \code{variable}
}
\description{
Filter a column of a list
}
\details{
This function allows you to filter a column of a list bases on a element of a variable within.
}
\author{
Eduardo Trujillo
}
