% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{firstvariables_df_datatype}
\alias{firstvariables_df_datatype}
\title{firstvariables_df_datatype}
\usage{
firstvariables_df_datatype(
  df,
  data_type = c("character", "numeric", "integer", "date", "numeric", "factor")
)
}
\arguments{
\item{df}{dataset to check.}

\item{data_type}{string datatype that we are interested.}
}
\value{
This function allows us to return:"
\itemize{
\item Names of the first or first, second or first, second and third variables that have the same data type.
\item In other case return a message indicate that the first variables do not have that data type.
}
}
\description{
Check if the first variables have an specific data type and return the names of that variables
}
\details{
This function allows you to check if the first or the first two or the first three variables have the same data type and return the names of that variables.
}
\author{
Eduardo Trujillo
}
