% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{footer_close_design}
\alias{footer_close_design}
\title{footer_close_design}
\usage{
footer_close_design(label, type_letter)
}
\arguments{
\item{label}{text to show}

\item{type_letter}{css type of letter}
}
\value{
This function returns a button with a beautiful text design
}
\description{
Create a button footer with a particular design
}
\details{
This function allows you to create a button footer with a particular design
}
\author{
Eduardo Trujillo
}
