% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{image_title}
\alias{image_title}
\title{image_title}
\usage{
image_title(src_image, width_image, height_image, style_image, TITLE)
}
\arguments{
\item{src_image}{url-link of an image to show in the title header}

\item{width_image}{width of the image}

\item{height_image}{height of the image}

\item{style_image}{csss argument}

\item{TITLE}{string of text to show in the title header}
}
\value{
return an HTML tag of the header bar
}
\description{
Create a title for the web page with an image
}
\details{
This function allows you to create a title with an image for your header
}
\author{
Eduardo Trujillo
}
