% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{improve_kmeans_labels}
\alias{improve_kmeans_labels}
\title{improve_kmeans_labels}
\usage{
improve_kmeans_labels(df, id, label, k)
}
\arguments{
\item{df}{dataset to change labels.}

\item{id}{dataset id variable reference of balance}

\item{label}{k-means label variable}

\item{k}{number of desire clusters}
}
\value{
desagregated dataset \code{df} with optimized K-labels
}
\description{
Optimize generated K-labels for desagregated dataset
}
\details{
\enumerate{
\item split the dataset by \code{id} testing unique elements on \code{label}.
Spliting on unique and duplicated sublist of the list improve_kmeans_labels
called \code{df_splited}.
\item For the duplicated sublist, we split it by the \code{label} testing
unique element on  \code{label} in unique elements and duplicated elements.
}
\itemize{
\item For the duplicated elements, then those are going to be the duplicated
sublist of 1.
\item For the unique elements, then those are going to be appended to the unique
sublist 1.
on each same specific sublist of the unique sublist 1.
\item Now we have a correct sublist of unique and duplicated elements.
}
\enumerate{
\item 3.From the duplicated sublist we take the first row of each sublist called
\emph{to_modify} and from the unique sublist we take a random sample of the
same length from the duplicated one called \emph{uniq_modify}. From that sublist
we create a sublist of the k-mean labels called \emph{uniq_labels}.
\item 4.We modify \emph{to_modify} based on the list of labels \emph{uniq_labels}
obtained from the sublist  \emph{uniq_modify} where if the label of to_modify
is in \emph{uniq_labels} then take a random number between 1 to k except that
labels of \emph{uniq_labels}. In other case take any label from the sublist of
the sublist \emph{uniq_labels}.
\item 5.The modify sublist \emph{to_modify} is going to be append in the list of
samples \emph{uniq_modify} in each sublist
\item 6.We modify the original created list \code{df_splited} modifying the
unique sublist elements with \emph{uniq_modify} sublist and modify the duplicated
sublist deleting the first row of each sublist since was used on 3.
\item 7.Create the original dataset with the modify labels.
\item 8.If the duplicated sublist still have duplicate elements the apply
recursively the function to change the label of thoss repeated.
}
}
\note{
This function is used to improve the k-means labels based on the id
variable.
}
\author{
Eduardo Trujillo
}
