% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{improve_kmeans_labels_variable}
\alias{improve_kmeans_labels_variable}
\title{improve_kmeans_labels_variable}
\usage{
improve_kmeans_labels_variable(
  df,
  id,
  label,
  var_model,
  split_type = c("mean_split", "range_split")
)
}
\arguments{
\item{df}{dataset to change labels.}

\item{id}{dataset id variable}

\item{label}{k-means label variable}

\item{var_model}{reference variable to balance}

\item{split_type}{type of label modification.}
}
\value{
This function modify the \code{label} variable of the \code{df} based
on \code{split_type}:
\itemize{
\item If \code{split_type == "mean_split"}, change the labels where are
upper the the mean of \emph{1.} to the lower ones, to balance
the grouped sum \code{var_model} by the \code{label}
\item If \code{split_type == "range_split"}, change the labels out of the
range population mean +- 1 standard deviation of \emph{1.}. Channging
equally to the previous option.
}
}
\description{
Improve the dataset labels based on a balanced grouped sum of a particular
variable.
}
\details{
\enumerate{
\item Get the grouped sum of  \code{var_model} by the created k-mean
\code{label} variable.
\item Calculate the population mean and standard deviation from \strong{1.} as
parameters to modify the \code{label}
\item create the \emph{value_check} as the constant values as comparation reference
where the user will select based on the parameter \code{split_type}
\item With help of the function \emph{split_lower_upper_df} we obtain the labels
under and upper the constant values on \strong{3.}
\item Apply the subfunction \emph{modify_labels_variable_df}
}
}
\note{
This function is used to improve the k-means labels based on a
particular variable.
}
\author{
Eduardo Trujillo
}
