% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{interactive_panelOutput}
\alias{interactive_panelOutput}
\title{interactive_panelOutput}
\usage{
interactive_panelOutput(id, outputID, ID_shinyJS)
}
\arguments{
\item{id}{module id}

\item{outputID}{output for the action Button}

\item{ID_shinyJS}{id of the body to hide and show}
}
\value{
return the server side of the interactive_panel function
}
\description{
Server part of the interactive_panel function
}
\details{
This function allows you to make interactively your interactive_panel function
}
\note{
Use this function at the same time with the function interactive_panel.
}
\author{
Eduardo Trujillo
}
