% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{intermediate_bar}
\alias{intermediate_bar}
\title{intermediate_bar}
\usage{
intermediate_bar(image_href, size_text, text)
}
\arguments{
\item{image_href}{url-link of an image to show in the jumbotron}

\item{size_text}{size of the text inside the bar}

\item{text}{string text or tag text you want to share.}
}
\value{
return a personalized bar with text that you want to share
}
\description{
Create an intermidiate personalized bar
}
\details{
This function allows you to make an horizontal bar in your webpage
}
\author{
Eduardo Trujillo
}
