% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{intermediate_bar_icon}
\alias{intermediate_bar_icon}
\title{intermediate_bar_icon}
\usage{
intermediate_bar_icon(
  title,
  title_text,
  icon_name,
  icon_class,
  icon_title,
  icon_text_reference
)
}
\arguments{
\item{title}{title of the information inside the jumbotron}

\item{title_text}{referenced text of the jumbotron}

\item{icon_name}{3 names of icons.}

\item{icon_class}{3 classes of icons.}

\item{icon_title}{3 titles which refer to the information that describes the icon.}

\item{icon_text_reference}{3 texts which refer to the information that describes the icon.}
}
\value{
return a personalized jumbotron with referenced text and a bar with 3 icons information.
}
\description{
Create an intermidiate personalized bar of icons
}
\details{
This function allows you to make a jumbotron and a bar of icons
}
\author{
Eduardo Trujillo
}
