% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval_between}
\alias{interval_between}
\title{interval_between}
\usage{
interval_between(x, rng)
}
\arguments{
\item{x}{vector}

\item{rng}{range of dimension 2}
}
\value{
Vector of TRUE/FALSE values of where a vector \code{x} is in that particular \code{rng}
}
\description{
Visualize a close interval from a vector.
}
\details{
This function allows you to identify a close interval.
}
\author{
Eduardo Trujillo
}
