% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{iterative_actionButton}
\alias{iterative_actionButton}
\title{iterative_actionButton}
\usage{
iterative_actionButton(
  id,
  inputIds,
  labels,
  icons,
  icons_lib,
  style = "font-family:system-ui;"
)
}
\arguments{
\item{id}{module id}

\item{inputIds}{vector of inputId of each actionButton}

\item{labels}{vector of label of each actionButton}

\item{icons}{vector of icon-name of each actionButton}

\item{icons_lib}{vector of library of each icon of each actionButton}

\item{style}{CSS modification of the action buttons}
}
\value{
return multiple action buttons without mentioning \code{actionButton} many times.
}
\description{
Create multiple action buttons
}
\details{
This function allows you to create multiple action buttons inside a shiny module
}
\author{
Eduardo Trujillo
}
