% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{iterative_merge}
\alias{iterative_merge}
\title{iterative_merge}
\usage{
iterative_merge(dfs_list, key, ...)
}
\arguments{
\item{dfs_list}{list of datasets}

\item{key}{join ID}

\item{...}{additional parameters of merge}
}
\value{
This function returns a merged dataset of all the datasets from \code{dfs_list}
}
\description{
Merge a list of datasets
}
\details{
This function allows you to make any type of merge of different datasets withing a list.
}
\author{
Eduardo Trujillo
}
