% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kmean_samesize}
\alias{kmean_samesize}
\title{kmean_samesize}
\usage{
kmean_samesize(df, k, nstart = 25)
}
\arguments{
\item{df}{dataset to generate k-means labels}

\item{k}{number of desire clusters}

\item{nstart}{number of k-mean iterations for the centroids creation.}
}
\value{
balanced k-means label for each observation of the dataset
}
\description{
Create optimal k-mean labels for a dataset.
}
\details{
\enumerate{
\item Calculate the k-means algorithm on numeric or integar variables from the
dataset \code{df}.
\item Calculate the euclidean distance between the row df points and each
centroid.
}
\itemize{
\item (x0,y0,z0,..),(x1,y1,z1,...),..,(xk,yk,zk,...) k centroids.
\item We are going to have k df s. Therefore, we use \emph{cbind.data.frame} to create
a df where each df is a variable with length of the original df.
}
\enumerate{
\item 3.variables used:
}
\itemize{
\item cardinality_sample: number of elements that are going to be for each
cluster (label).
\item ctrl_clstr_no_elmnts: this variable helps to control the balance
cluster size.
\item label: result balanced label that is going to be created.
}
\enumerate{
\item 4.For each iteration (each df row):
\item 4.1. We obtain the row index where is the minimum euclidean distance in that
row of the df 2. defined as bestcluster and in that index we add 1 to the
observation of ctrl_clstr_no_elmnts variable representing adding 1 intended
top to k.
}
\itemize{
\item example:
}
\itemize{
\item iteration 1 - bestcluster = 2
\item ctrl_clstr_no_elmnts  = 0 1 0 0 0 0 | row 1 of dataset \emph{euclidean_distance}
\item iteration 2 - bestcluster = 2
\item ctrl_clstr_no_elmnts  = 0 2 0 0 0 0 | row 2 of dataset \emph{euclidean_distance}
\item iteration 3 - bestcluster = 3
\item ctrl_clstr_no_elmnts  = 0 2 1 0 0 0 | row 3 of dataset \emph{euclidean_distance}
\item iteration 4 - bestcluster = 2
\item ctrl_clstr_no_elmnts  = 0 3 1 0 0 0 | row 4 of dataset \emph{euclidean_distance}
}
\enumerate{
\item 4.2.Add if statement to control and balance the size of each cluster based on
ctrl_clstr_no_elmnts variable. If in the index bestcluster of the observation
of ctrl_clstr_no_elmnts is greater than cardinality_sample then define the
variable on index bestcluster of df2. as NA to not affect the next iteration
and have only the columns in euclidean_distance where ctrl_clstr_no_elmnts is
less than cardinality_sample
}
}
\note{
\itemize{
\item The kmeans algorithm assigns a label for each observation. However,
it is not necessary to have balanced clusters in the amount of
observations that is why we use this function.
\item Kmeans algorithm assign randomly the centroids based on the
iterations of \code{nstart}. This is why every time this
function runs the output would have a different order in the labels.
\item It is recommendable that the size of the population is divisible by
the number of clusters \code{k} to have equal number of elements in
each cluster.
}
}
\author{
Eduardo Trujillo
}
