% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_of_lists}
\alias{list_of_lists}
\title{list_of_lists}
\usage{
list_of_lists(no_sublists, element_sublists)
}
\arguments{
\item{no_sublists}{number of sublists}

\item{element_sublists}{element to instert in each sublist}
}
\value{
Using \code{no_sublists} and \code{element_sublists} list, which each sublist contains the same element
}
\description{
Create a list of sublists with the same elements each
}
\details{
This function allows you to create a list where you can add in each sublist the same element
}
\note{
Add this function inside a UI in your shiny app.
}
\author{
Eduardo Trujillo
}
