% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{loginOutput}
\alias{loginOutput}
\title{loginOutput}
\usage{
loginOutput(id, user_base, username, password, waiter = waiting_screen)
}
\arguments{
\item{id}{module id}

\item{user_base}{dataset that contain the users and the passwords to log-in}

\item{username}{variable of the \code{user_base} dataset that represent the users}

\item{password}{variable of the \code{user_base} dataset that represent the passwords}

\item{waiter}{created waiter to use for the load part}
}
\value{
return the server side of the function \emph{loginUI}
}
\description{
Server part of the \emph{loginUI} function
}
\details{
This function allows you to make the server work of the \emph{loginUI} function
}
\note{
Use this function at the same time with the function \emph{loginUI}.
}
\author{
Eduardo Trujillo
}
