% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{loginUI}
\alias{loginUI}
\title{loginUI}
\usage{
loginUI(
  id,
  image_href,
  title,
  error_message,
  company_name,
  header = NULL,
  sidebar = NULL,
  body = NULL,
  rightsidebar = NULL
)
}
\arguments{
\item{id}{module id}

\item{image_href}{url-link of an image to show in the log-in pannel and the load part}

\item{title}{string of text to show in the title header}

\item{error_message}{message string to show when your password is incorrect}

\item{company_name}{name of your company to show in the footer sign up and in the load part}

\item{header}{header of the dashboardPagePlus}

\item{sidebar}{sidebar of the dashboardPagePlus}

\item{body}{body of the dashboardPagePlus}

\item{rightsidebar}{rightsidebar of the dashboardPagePlus}
}
\value{
"This function returns \code{two possible UI}:"
\itemize{
\item If header & sidebar & body & rightsidebar are \emph{NULL} then return a created UI
\item If header & sidebar & body & rightsidebar are not \emph{NULL} then return a \emph{dashboardPagePlus}
}
}
\description{
Create an personalized log in to a web page
}
\details{
This function allows you to create an interactive personalized log-in webpage
}
\note{
Use this function at the same time with the function \emph{loginOutput}.
}
\author{
Eduardo Trujillo
}
