% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{marg_dist}
\alias{marg_dist}
\title{marg_dist}
\usage{
marg_dist(df, num_int_var)
}
\arguments{
\item{df}{dataset to obtain the descriptive statistical analysis}

\item{num_int_var}{\emph{ONE} NUMERICAL VARIABLE from the dataset}
}
\value{
List where each element is the marginal distribution of \code{num_int_var} from each combination levels of the factor variables.
}
\description{
Create the marginal distribution of each combination of levels from the factor variables.
}
\details{
This function allows you to create the marginal distribution of a numeric variable for each combination of levels from the factor variables.
}
\note{
\itemize{
\item If there is no factor variable(s) then it is going to bring the whole distribution of  \code{num_int_var}
\item This function is applied to one numeric variable \code{num_int_var}
}
}
\author{
Eduardo Trujillo
}
