% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{message_reactive}
\alias{message_reactive}
\title{message_reactive}
\usage{
message_reactive(id, text, type = c("success", "fail"))
}
\arguments{
\item{id}{module id}

\item{text}{string to show for success or fail}

\item{type}{type of box message}
}
\value{
"This function returns \code{two possible box messages}:"
\itemize{
\item If \code{type} is \code{success} then returns success box message
\item If \code{type} is \code{fail} then returns fail box message
}
}
\description{
Create an interactive box message
}
\details{
This function allows you to create an interactive personalized success or fail message
}
\author{
Eduardo Trujillo
}
