% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{modify_labels_variable_df}
\alias{modify_labels_variable_df}
\title{modify_labels_variable_df}
\usage{
modify_labels_variable_df(
  df,
  id,
  label,
  var_model,
  check,
  check_mean,
  check_labels,
  sum_var_model
)
}
\arguments{
\item{df}{dataset to change labels.}

\item{id}{dataset id variable}

\item{label}{k-means label variable}

\item{var_model}{reference variable to balance}

\item{check}{grouped sum from improve_kmeans_labels_variable function}

\item{check_mean}{population mean from
improve_kmeans_labels_variable function}

\item{check_labels}{labels under and upper the constant values from
improve_kmeans_labels_variable function}

\item{sum_var_model}{variable from check that represent the grouped sum from
improve_kmeans_labels_variable function}
}
\value{
The provided dataset \code{df} modifying the \code{label} variable.
}
\description{
Improve the dataset labels based on a balanced grouped sum of a particular
variable.
}
\details{
\enumerate{
\item Based on the separeted labels (4. of improve_kmeans_labels_variable)
\code{check_labels} for the uppers get the difference between the
observations of \code{sum_var_model} in each level on \code{check}
and the mean (total sum of each level where is necessary to quit from
upper and pass to under).
\item From the previous If we have only one total sum to quit from upper
then return that list in other case get the maximums according as the number
of unders as well as minimums. For the previous get a maximum and minimum for
each under (uppers that are going to pass to the uders)
\item From the previous set the name of each sublist as the names of the unders
from the list \code{check_labels} to know which is going to pass from upper
to which from unders.
\item Modify the dataset, based on the previous each under (max and min) change
the label of the uppers to unders to balance the dataset.
}
}
\note{
This function is a \emph{subfunction} of \strong{improve_kmeans_labels_variable}
}
\seealso{
\href{https://stackoverflow.com/questions/45231735/match-values-to-nearest-value-in-another-array-in-r}{nearest values}
}
\author{
Eduardo Trujillo
}
