% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{mongo_Listread_or_write}
\alias{mongo_Listread_or_write}
\title{mongo_Listread_or_write}
\usage{
mongo_Listread_or_write(
  collection = config$collection,
  database = config$database,
  host = config$host,
  username = config$username,
  password = config$password,
  do = c("read", "write"),
  json_to_write = NULL
)
}
\arguments{
\item{collection}{your cluster}

\item{database}{name of database inside the \code{cluster}}

\item{host}{host of the \code{cluster}}

\item{username}{Your username to enter to the \code{cluster}}

\item{password}{your password to enter to the \code{cluster}}

\item{do}{\code{read} database in the cluster OR \code{write} information in the database}

\item{json_to_write}{if \code{do} is \code{write} then this parameter represent the information}
}
\value{
This function returns two possible results depend on \code{do} parameter:
\itemize{
if \code{do} = read then we return the database information as a list as a global variable
if \code{do} = write then write a JSON file (or other) to the database
}
}
\description{
Read or pull data in a cluster database
}
\details{
This function allows you to pull a json file to the database o read information in the database
}
\note{
Add this function inside a UI in your shiny app.
}
\author{
Eduardo Trujillo
}
