% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{mongo_manipulation}
\alias{mongo_manipulation}
\title{mongo_manipulation}
\usage{
mongo_manipulation(
  collection = config$collection,
  database = config$database,
  host = config$host,
  username = config$username,
  password = config$password,
  mongo_choice = c("push", "pull", "update_record", "delete_record"),
  push_record = NULL,
  id_filter = NULL,
  element_id_filter = NULL,
  id_update = NULL,
  element_id_update = NULL
)
}
\arguments{
\item{collection}{your cluster}

\item{database}{name of database inside the \code{cluster}}

\item{host}{host of the \code{cluster}}

\item{username}{Your username to enter to the \code{cluster}}

\item{password}{your password to enter to the \code{cluster}}

\item{mongo_choice}{choice that we want to do with a database in a collection}

\item{push_record}{dataset we want to push to the collection}

\item{id_filter}{variable of the dataset that we want to filter when we pull this.}

\item{element_id_filter}{observation of \code{id_variable} that will allow to pull an specific table.}

\item{id_update}{variable of the dataset that we want to update.}

\item{element_id_update}{values to update for the variable \code{id_update}}
}
\value{
#' "This function \emph{different results} based on \code{mongo_choice} variable:"
\itemize{
\item If \code{mongo_choice = "push"} then you only need to use the variable \code{push_record} to push the table with the information to your MongoDB collection.
\item If \code{mongo_choice = "pull"} is not necessary to use other variable, because you are going to pull the dataset from the config file.
\item If \code{mongo_choice = "pull"} and use the variables \code{id_filter} and \code{element_id_filter} to pull a table from the dataset.
\item If \code{mongo_choice = "update_record"} and use the variables \code{id_filter}, \code{element_id_filter}, \code{id_update}, \code{element_id_update} to update a record from a table from the dataset.
\item If \code{mongo_choice = "delete_record"} and use the variables \code{id_filter} and \code{element_id_filter} to delete a table from the dataset.
}
}
\description{
Read or pull or modify data in a cluster database
}
\details{
This function allows you to push, pull or modify a database
}
\note{
You can use this function inside your shiny app.
}
\author{
Eduardo Trujillo
}
