% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{nas_before_after_outliers}
\alias{nas_before_after_outliers}
\title{nas_before_after_outliers}
\usage{
nas_before_after_outliers(df, factor_vars = NULL)
}
\arguments{
\item{df}{dataset that we want to clean outliers}

\item{factor_vars}{factor variables to delete outliers}
}
\value{
This function allows us to return a list which each sublist is for each of the integer and numeric variables where is going to be a dataset that counts the number of NAs before and after applying the function \emph{delete_outliers} based on the factor variables (if there is one).
}
\description{
Count the number of NAs before and after applying the function \emph{delete_outliers}
}
\details{
This function allows you to have a list where each sublist is the difference of NAs for each integer or numeric variable.
}
\author{
Eduardo Trujillo
}
