% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{nav_tabPanel}
\alias{nav_tabPanel}
\title{nav_tabPanel}
\usage{
nav_tabPanel(id, title, outputID, TITLE)
}
\arguments{
\item{id}{module id}

\item{title}{vector of the tabs inside the navbarMenu}

\item{outputID}{vector of outputs for the renderUI}

\item{TITLE}{title for the navbarMenu}
}
\value{
return a navbarMenu with tabs where you can use for create UI elements
}
\description{
Create iterative elements inside a navbarMenu
}
\details{
This function allows you to create more than one tabpanel inside a navbarMenu dynamically
for use inside a navbarpage
}
\note{
Add this function inside a navbar
}
\author{
Eduardo Trujillo
}
