% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{NotificationmenuInput}
\alias{NotificationmenuInput}
\title{NotificationmenuInput}
\usage{
NotificationmenuInput(id)
}
\arguments{
\item{id}{module id}
}
\value{
UI-message component
}
\description{
Create a UI-Message component
}
\details{
This function allows you to create a UI-message component inside a shiny module
}
\note{
Use this function with function \emph{NotificationmenuOutput}.
}
\author{
Eduardo Trujillo
}
