% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{NotificationmenuOutput}
\alias{NotificationmenuOutput}
\title{NotificationmenuInput}
\usage{
NotificationmenuOutput(id, message)
}
\arguments{
\item{id}{module id}

\item{message}{list of message we want to share}
}
\value{
UI-message component
}
\description{
Create server side of the UI-message
}
\details{
This function allows you to make work the UI-message component inside a shiny module
}
\note{
Use this function with function \emph{NotificationmenuInput}.

"This message list needs to have the variables:"
\itemize{
\item \code{text} refers to the messages we want to share
\item \code{status} color of the messages ("warning", "success", "info", etc.)
\item \code{Href} references link
\item \code{icon} icon of the message
}
}
\author{
Eduardo Trujillo
}
