% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{obtain_regex}
\alias{obtain_regex}
\title{obtain_regex}
\usage{
obtain_regex(
  pattern = NULL,
  return_regex = c("starts_with_pattern", "not_starts_with_pattern",
    "ends_with_pattern", "not_ends_with_pattern", "before_pattern", "after_pattern",
    "between_pattern", "contains_pattern", "not_contains_pattern", "extract_pattern",
    "or", "special_chrs", "regex_dates")
)
}
\arguments{
\item{pattern}{based on this string we obtain the regex.}

\item{return_regex}{choice of regex you want to obtain.}
}
\value{
This function return \emph{different results} based on \code{return_regex} argument.
}
\description{
Obtain a specific regex to apply in a vector or column
}
\details{
This function allows you to obtain a regex based on your necesities with the intention that you do not focus on the regex part.
}
\note{
\itemize{
\item If \code{return_regex = "or"}, then \code{pattern} could be a vector of length greater or equal to 1.
\item If \code{return_regex = "between_pattern"}, then \code{pattern} needs to be a vector of length 2.
\item If \code{return_regex = "special_chrs"} or \code{return_regex = "regex_dates"}, then we do not need to specify \code{pattern}, since we are going to return a vector of specific regexes.
\item Other case the \code{pattern} needs to have length greater or equal to 1.
\item If you want to get the regex inside another regex you need to apply more than once this function, since the output is only based on \code{pattern}
}
}
\author{
Eduardo Trujillo
}
