% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{passOutput}
\alias{passOutput}
\title{passOutput}
\usage{
passOutput(id)
}
\arguments{
\item{id}{module id}
}
\value{
return the server side of the function \emph{passUI}
}
\description{
Server part of the \emph{passUI} function
}
\details{
This function allows you to make the server work of the \emph{passUI} function
}
\note{
\itemize{
\item Use this function at the same time with the function \emph{passUI}.
\item his function returns the password and the confirmation if we need it later (as a reactive list).
}
}
\author{
Eduardo Trujillo
}
