% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{passUI}
\alias{passUI}
\title{passUI}
\usage{
passUI(id)
}
\arguments{
\item{id}{module id}
}
\value{
passwords and its confirmation password and a click button to show what we are writting on them.
}
\description{
Create a dynamic password visualization
}
\details{
This function allows you to visualize the string password and its confirmation based on a click
}
\note{
Use this function at the same time with the function \emph{passOutput}.
}
\author{
Eduardo Trujillo
}
