% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{personalize_jumbotron}
\alias{personalize_jumbotron}
\title{personalize_jumbotron}
\usage{
personalize_jumbotron(
  image_href,
  width_text,
  heigth_text,
  header_text,
  text,
  cited,
  href
)
}
\arguments{
\item{image_href}{url-link of an image to show in the jumbotron}

\item{width_text}{width of the text square}

\item{heigth_text}{width of the text square}

\item{header_text}{title of the text square}

\item{text}{the string of a phrase you want to share}

\item{cited}{who made the text quote}

\item{href}{link to open as a reference of the quote}
}
\value{
return a particular jumbotron with a quote insie of a text square
}
\description{
Create a personalized jumbotron
}
\details{
This function allows you to make a particular example of jumbotron
}
\author{
Eduardo Trujillo
}
