% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{pull_from_environment}
\alias{pull_from_environment}
\title{pull_from_environment}
\usage{
pull_from_environment(config_file, id_filter = NULL, element_id_filter = NULL)
}
\arguments{
\item{config_file}{config file that confain the database information}

\item{id_filter}{variable of the dataset that we want to filter when we pull this.}

\item{element_id_filter}{observation of \code{id_variable} that will allow to pull an specific table.}
}
\value{
"This function allows you to pull dataset without overwriting the previous pulled."
\itemize{
\item If \code{mongo_choice = "pull"} is not necessary to use other variable, because you are going to pull the dataset from the config file.
\item If \code{mongo_choice = "pull"} and use the variables \code{id_filter} and \code{element_id_filter} to pull a table from the dataset.
}
}
\description{
particular pull data
}
\details{
This function allows you to pull a database similar with the function \emph{mongo_manipulation}
}
\note{
\itemize{
\item You can use this function inside your shiny app.
\item Since the function  \emph{mongo_manipulation} pull a variable \code{pull_info} in the global environment, and if we want to pull from different \emph{cofig} files. Then use this functions to not overwrite the variable \code{pull_info}, since now is written in son environment.
}
}
\author{
Eduardo Trujillo
}
