% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{RDs_and_S3}
\alias{RDs_and_S3}
\title{RDs_and_S3}
\usage{
RDs_and_S3(
  dfs = NULL,
  objects = NULL,
  bucket,
  choice = c("watch_bucket", "push", "pull", "delete")
)
}
\arguments{
\item{dfs}{dataset or a list of datasets}

\item{objects}{vector of files in S3 (it could be folder/file path)}

\item{bucket}{name of your S3 bucket}

\item{choice}{choice that we want to do with the RDs data}
}
\value{
"This function returns \emph{different results} based on \code{choice} variable:"
\itemize{
\item If \code{choice = "watch_bucket"} then you only need to use the variable \code{bucket} to see what is inside the AWS S3 bucket.
\item If \code{choice = "push"} then you need to use the variable \code{dfs} representing the dataset (or a list of datasets) in your R session, the  \code{objects} representing the path on AWS S3. and the name of your \code{bucket}.
\item If \code{choice = "pull"} then you need to use  the variable \code{objects} representing a vector of path of your RDs data on AWS S3. and the name of your \code{bucket} to pull in our R-session.
\item If \code{choice = "delete"} then you need to use  the variable \code{objects} representing a vector of path of your RDs data on AWS S3. and the name of your \code{bucket} to delete in AWS S3.
}
}
\description{
Read an S3 bucket or push, pull or delete an RDs data from the S3 bucket.
}
\details{
This function allows you to read an S3 bucket or manipulate an RDs data with S3 AWS
}
\note{
\itemize{
\item You can use this function inside your shiny app.
\item If \code{mongo_choice = "push"} then in AWS S3 is going to be an RDs data (s).
}
}
\author{
Eduardo Trujillo
}
