% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(path, encoding = "UTF-8", sheet = 1, col_select = NULL, ...)
}
\arguments{
\item{path}{path where is your dataset in your local directory}

\item{encoding}{type of code form}

\item{sheet}{page you want to import if you are pulling a \code{.xls, .xlsx} file}

\item{col_select}{vector of specific variables from your dataset that you want to use (rather than the all dataset).}

\item{...}{additional arguments to use to import the dataset.}
}
\value{
"This function allows us to \emph{pull a clean dataset} with great characteristics:"
\itemize{
\item Use any of this format: \code{.sas7bdat, .txt, .csv, .rds, .xls, .xlsx}
\item Use the parameter \code{col_select}, to select variables that we want from the dataset
\item \code{Delete special characters}
\item \code{Toupper Variables}
\item \code{Toupper character and factor observations}
\item \code{Assign the correct data type to each variable}
\item \code{Change "" strings to N/A}
\item \code{Remove whitespace from start and end of each observation}
}
}
\description{
Pull a dataset from your own directory files.
}
\details{
This function allows you to pull a dataset with many types of formats.
}
\author{
Eduardo Trujillo
}
