% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{refernce_icon}
\alias{refernce_icon}
\title{refernce_icon}
\usage{
refernce_icon(image_href, href, width_image, width_column)
}
\arguments{
\item{image_href}{url-link of the image}

\item{href}{url-link of the reference web page of the icon}

\item{width_image}{width of the image}

\item{width_column}{width of the column in the fluidRow}
}
\value{
return an image that refers to an URL link.
}
\description{
Create a reference shadow image.
}
\details{
This function allows you to create an image with gray format that links to a web page.
}
\author{
Eduardo Trujillo
}
