% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report_NumInVar}
\alias{report_NumInVar}
\title{report_NumInVar}
\usage{
report_NumInVar(day, df, date_variable, num_int_var, factor_variable)
}
\arguments{
\item{day}{vector of filter dates on the \code{date_variable}}

\item{df}{dataset to obtain the report}

\item{date_variable}{string that represent variable´s name of data type date}

\item{num_int_var}{string that represent variable´s name of data type numeric or integer}

\item{factor_variable}{string that represent variable´s name of data type factor}
}
\value{
This function return a list with two datasets selecting only specific \code{day}´s from the \code{date_variable}, where is going to be a summary of all the numeric and integer variables and other containing a grouped dataset
}
\description{
Create a summary and a grouped dataset from the orignal dataframe.
}
\details{
This function allows you to obtain a summary and a grouped dataset from the orignal dataframe
}
\note{
\itemize{
\item If the \code{day} is all the observations from \code{date_variable}, this means that we are creating historical summary and historical grouped dataset.
\item The variables created on the summary are N:= number of observations per group, SUM:= total of observations per group, PERCENT:= participation percentage of each group from each numeric and integer variable
}
}
\author{
Eduardo Trujillo
}
