% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{rightInputs_Navbar}
\alias{rightInputs_Navbar}
\title{rightInputs_Navbar}
\usage{
rightInputs_Navbar(..., inputs)
}
\arguments{
\item{...}{elements of the navbarPage}

\item{inputs}{inputs to use in your navbar}
}
\value{
Create a navbarPage using \code{inputs} in the right side
}
\description{
Create a navbar allowing inputs in the right side
}
\details{
This function allows you to create a standard navarPage, with the possibility to
add inputs in the right side of the bar
}
\note{
Use this function in the same way as a Navar Page
}
\author{
Eduardo Trujillo
}
