% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{semester_quarter}
\alias{semester_quarter}
\title{semester_quarter}
\usage{
semester_quarter(df, year, previous = FALSE, semester = NULL, quarter = NULL)
}
\arguments{
\item{df}{dataset to obtain the dates}

\item{year}{year that we want to obtain the vector}

\item{previous}{If we want to obtain the semester or quarter from the previous year}

\item{semester}{Number of the semester that we want to obtain the vector 1:2}

\item{quarter}{Number of the quarter that we want to obtain the vector 1:4}
}
\value{
"This function returns \emph{different results} based on the arguments \code{select_year}, \code{select_month} & \code{previous}".
\itemize{
\item If \code{previous = FALSE} and \code{semester} return a vector of observations of that semester of the year \code{select_year} in the date variable.
\item If \code{previous = TRUE} and  \code{semester} return a vector of observations of that semester of the previous year \code{select_year} in the date variable.
\item If \code{previous = FALSE} and \code{quarter} return a vector of observations of that quarter of the year \code{select_year} in the date variable.
\item If \code{previous = TRUE} and  \code{quarter} return a vector of observations of that quarter of the previous year \code{select_year} in the date variable.
}
}
\description{
Return a vector of possible semesters or quarters in a year.
}
\details{
This function allows you to return a vector of observations from a semester or quarter of a year (or the previous year)
}
\note{
\itemize{
\item If \code{semester} and \code{quarter} are selected both, then return a string message.
\item If there is something wrong with the \code{year} and/or \code{previous} option defined in the firm, then return a string message.
\item If there is no dates in the \code{semester} or \code{quarter} defined, then return a string message.
}
}
\author{
Eduardo Trujillo
}
