% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{sign_upOutput}
\alias{sign_upOutput}
\title{sign_upOutput}
\usage{
sign_upOutput(id, pass, confirm_pass, config)
}
\arguments{
\item{id}{module id}

\item{pass}{reactive password obtained from the function \code{passOutput}}

\item{confirm_pass}{reactive confirmation password obtained from the function \code{passOutput}}

\item{config}{\code{YAML} file indicate the mongodb database where we are going to push the new users}
}
\value{
return the server side of the function \emph{sign_upUI}
}
\description{
Server part of the \emph{sign_upUI} function
}
\details{
This function allows you to make the server work of the \emph{sign_upUI} function
}
\note{
\itemize{
\item Use this function at the same time with the function \emph{sign_upUI}.
\item The output of this function is a data.table that is going to push the recollected information to a MongoDB. So you need to have a collection.
}
}
\author{
Eduardo Trujillo
}
