% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{sign_upUI}
\alias{sign_upUI}
\title{sign_upUI}
\usage{
sign_upUI(id, image_href, title, services, login_href)
}
\arguments{
\item{id}{module id}

\item{image_href}{url-link of an image of your company}

\item{title}{your company name as string}

\item{services}{list of services your company offer.}

\item{login_href}{url-link reference to go to another page (the log-in web page)}
}
\value{
UI webpage of the sign up.
}
\description{
Create an personalized sign up web page
}
\details{
This function allows you to create an interactive personalized sign up webpage
}
\note{
Use this function at the same time with the function \emph{sign_upOutput}.
}
\author{
Eduardo Trujillo
}
