% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_lower_upper_df}
\alias{split_lower_upper_df}
\title{split_lower_upper_df}
\usage{
split_lower_upper_df(df, vars, value_check, names)
}
\arguments{
\item{df}{dataset to split}

\item{vars}{variable reference to split the dataset}

\item{value_check}{constant value as comparation reference}

\item{names}{of the sublists}
}
\value{
list with two sublists:
\itemize{
\item A sublist with the dataset filter for values of the
variable \code{vars} under \code{value_check}
\item A sublist with the dataset filter for values of the
variable \code{vars} upper \code{value_check}
}
}
\description{
Dataset comparation split
}
\details{
Split a dataset based on a comparison between values of a variable and a
constant value.
}
\author{
Eduardo Trujillo
}
