% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull&push.R
\name{sql_manipulation}
\alias{sql_manipulation}
\title{sql_manipulation}
\usage{
sql_manipulation(
  dsn = NULL,
  server = NULL,
  database,
  uid,
  pwd,
  sql_choice = c("push", "pull", "tables_info"),
  df_identifier = NULL,
  table_name = NULL,
  df = NULL,
  attributes = NULL
)
}
\arguments{
\item{dsn}{our \code{server} if is local}

\item{server}{our \code{server} if is not local}

\item{database}{database in the \code{server}}

\item{uid}{Your username to enter to the \code{server}}

\item{pwd}{your password to enter to the \code{server}}

\item{sql_choice}{choice that we want to do with a table in a database}

\item{df_identifier}{string name that identify a table in a database.}

\item{table_name}{name of the table}

\item{df}{dataset that we want to pull to the database}

\item{attributes}{variables to select from the table}
}
\value{
"This function returns \emph{different results} based on \code{mongo_choice} variable:"
\itemize{
\item If \code{sql_choice = "push"} then you need to use the parameters  \code{df_identifier}, \code{table_name}, \code{df} and if you want to select specific variables tu push in the table use the parameter \code{attributes}. All of these to push a table to the database.
\item If \code{sql_choice = "pull"} then  you need to use the parameters  \code{df_identifier} and  \code{table_name} (or use only \code{table_name} indicating the  \code{df_identifier} and  \code{table_name} together) and if you want to select specific variables tu pul from the table use the parameter \code{attributes}. All of these to pull to our R-studio session the table from the database.
\item If \code{sql_choice = "tables_info"} then you need to use the parameter \code{df_identifier} to bring a list of all the tables names (with its variables).
}
}
\description{
push or pull or watch the information of the database
}
\details{
This function allows you to push or pull or watch the information of the database
}
\note{
If there is something wrong with the connection will return a message string.
}
\author{
Eduardo Trujillo
}
