% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_equality}
\alias{test_equality}
\title{test_equality}
\usage{
test_equality(var)
}
\arguments{
\item{var}{vector variable}
}
\value{
This function returns \code{two possible results}:
\itemize{
\item \emph{TRUE} indicates that the vector has unique elements.
\item \emph{FALSE} indicates that the vector has repeated elements.
}
}
\description{
Check if elements of a variable are unique
}
\details{
This function has been design of a given vector
}
\author{
Eduardo Trujillo
}
