% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{title_bar}
\alias{title_bar}
\title{title_bar}
\usage{
title_bar(image_href, title)
}
\arguments{
\item{image_href}{url-link of an image to show in the title bar}

\item{title}{string of text to show in the title bar}
}
\value{
Using \code{image_href} and \code{title} return an HTML tag of the title bar
}
\description{
Create a title bar for the header of the webpage
}
\details{
This function allows you to put an image and a text in the title bar
of the web page
}
\note{
Add this function inside a UI in your shiny app.
}
\author{
Eduardo Trujillo
}
