% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udeploy.R
\docType{package}
\name{udeploy}
\alias{udeploy}
\alias{udeploy-package}
\title{udeploy: A package for deploy analysis and web-applications.}
\description{
This packages provides a variety of functions to generate automatic
statistics, analysis and web applications.
}
\author{
Eduardo Trujillo \email{act.jetv@gmail.com}
}
