% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{UIelement}
\alias{UIelement}
\title{UIelement}
\usage{
UIelement(id, outputID)
}
\arguments{
\item{id}{module id}

\item{outputID}{output for the renderUI}
}
\value{
Using \code{id} and \code{outputID} return a UI component
}
\description{
Create a UI component
}
\details{
This function allows you to create a UI component inside a shiny module
}
\note{
Use this function with a RenderUI.
}
\author{
Eduardo Trujillo
}
