% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{year_month}
\alias{year_month}
\title{year_month}
\usage{
year_month(df, select_year, select_month = NULL, previous = FALSE)
}
\arguments{
\item{df}{dataset to obtain the dates}

\item{select_year}{year that we want to obtain the vector}

\item{select_month}{name of the month of the year that we want to obtain the vector}

\item{previous}{If we want to obtain the previous year or month}
}
\value{
"This function returns \emph{different results} based on the arguments \code{select_year}, \code{select_month} & \code{previous}".
\itemize{
\item If \code{previous = FALSE} and  \code{select_year} return a vector of dates of that year in the date variable.
\item If \code{previous = TRUE} and  \code{select_year} return a vector of dates of the previous year in the date variable.
\item If \code{previous = FALSE} and  \code{select_month} return a vector of dates of that month of the year \code{select_year} in the date variable.
\item If \code{previous = TRUE} and  \code{select_month} return a vector of dates of the previous month year \code{select_year} in the date variable.
}
}
\description{
Return a vector of possible dates of a years, or dates of a month in a year.
}
\details{
This function allows you to return a vector of dates from a year (or the previous one) OR dates of a month of a year (or the previous month)
}
\note{
\itemize{
\item If the \code{df} does not have a date variable, then return a string message.
\item If the selected \code{select_year} is not included in the observations of the date variable from \code{df} then return a string message.
\item If \code{previous = TRUE} & \code{select_month = "JANUARY"} then will return a string message, since it returns the observations of the previous of the exact year \code{select_year} ( \emph{not the previous one} ).
\item If \code{previous = TRUE} and there is no previous \code{select_year} or \code{select_month}, then will return a string message.
}
}
\author{
Eduardo Trujillo
}
