% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{general_descript_stats}
\alias{general_descript_stats}
\title{general_descript_stats}
\usage{
general_descript_stats(
  df,
  num_int_var,
  statistical_choice = c("TABLES.DESCRIPTIVE_FREQUENCY", "TABLES.CENTRAL_TENDENCY",
    "TABLES.MODE", "TABLES.DISPERSION_TENDENCY",
    "TABLES.STATISTICAL_COEFFICIENTS.COEFFICIENTS",
    "TABLES.STATISTICAL_COEFFICIENTS.DESCRIPTIONS",
    "TABLES.CONDITIONAL.CONDITIONAL_FREQUENCY",
    "TABLES.CONDITIONAL.CONDITIONAL_EXPENCTATION", "PLOTS.DESCRIPTIVE_FREQUENCY.OGIVE",
    "PLOTS.DESCRIPTIVE_FREQUENCY.CUMMULATIVE", "PLOTS.DESCRIPTIVE_FREQUENCY.HISTOGRAM",
    "PLOTS.CENTRAL_TENDENCY.GENERAL_BOXPLOT", "PLOTS.CENTRAL_TENDENCY.BINS_BOXPLOT",
    "PLOTS.STANDARD_ERROR",      "PLOTS.COMPARING_MEANS")
)
}
\arguments{
\item{df}{dataset to obtain the descriptive statistical analysis}

\item{num_int_var}{\emph{ONE} NUMERICAL VARIABLE from the dataset}

\item{statistical_choice}{Type of descriptive statistical analysis}
}
\value{
"This function returns \emph{different results} based on the arguments \code{statistical_choice} argument".
\itemize{
\item If \code{statistical_choice = "TABLES.DESCRIPTIVE_FREQUENCY"}, list where each element is going to be the \emph{frequency table} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.CENTRAL_TENDENCY"},list where each element is going to be a \emph{dataset with the mode, mean min, q25, median, q75,p90,p95,p99,max, midrange} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.MODE"}, list where each element is going to be an \emph{interpretation of the mode} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.DISPERSION_TENDENCY"},list where each element is going to be a \emph{dataset with the variance, standard deviation, interquantil range and range} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.STATISTICAL_COEFFICIENTS.COEFFICIENTS"}, list where each element is going to be a \emph{list with the coefficient of variation, standard error, skeweness and kurtosis} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.STATISTICAL_COEFFICIENTS.DESCRIPTIONS"}, list where each element is going to be a \emph{list with the interpretation of the coefficient of variation, standard error, skeweness and kurtosis} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.CONDITIONAL.CONDITIONAL_FREQUENCY"}, list where each element is going to be a \emph{dataset with the conditional probability on that specific level-combination given the join observations on other specific level-combination} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "TABLES.CONDITIONAL.CONDITIONAL_EXPENCTATION"}, list where each element is going to be a \emph{dataset with the conditional expectancy, conditional variance & conditional standard deviation on that specific level-combination given the join observations on other specific level-combination} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "PLOTS.DESCRIPTIVE_FREQUENCY.OGIVE"}, list where each element is going to be the \emph{ogive plot} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "PLOTS.DESCRIPTIVE_FREQUENCY.CUMMULATIVE"}, list where each element is going to be the \emph{accumulative distribution function plot} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "PLOTS.DESCRIPTIVE_FREQUENCY.HISTOGRAM"}, list where each element is going to be the \emph{histogram plot} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "PLOTS.CENTRAL_TENDENCY.GENERAL_BOXPLOT"}, list where each element is going to be the \emph{box plot} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "PLOTS.CENTRAL_TENDENCY.BINS_BOXPLOT"}, list where each element is going to be the \emph{box plot on each bin from the frequency table} for the distribution of the \code{num_int_var} on a specific combination level of the factor variables
\item If \code{statistical_choice = "PLOTS.STANDARD_ERROR"}, \emph{bar plot} comparing the \emph{standard error} of each combination level of the factor variables for the distribution of the \code{num_int_var}.
\item If \code{statistical_choice = "PLOTS.COMPARING_MEANS"}, \emph{Polar area plot} comparing the \emph{mean} of each combination level of the factor variables for the distribution of the \code{num_int_var}.
}
}
\description{
Create a descriptive analysis from a numeric variable based on the levels of factor variables from a dataset.
}
\details{
This function allows you to return different types descriptive statistic analysis from a numeric variable based on the levels of factor variables from a dataset.
}
\note{
\itemize{
\item Each output is going to be a list, where each element is the result from each combination levels of the factor variables.
\item It is necessary that the dataset \code{df} contains at least one factor variable.
}
}
\author{
Eduardo Trujillo
}
